Handle server/network disconnects during an SMTP session
Previously the SMTPClientFutureListener.operationComplete method was
never being called if the server unexpectedly disconnected, such as might
happen if the server crashes after receiving DATA.

Added a Netty upstream handler that throws a new SMTPDisconnectedException
when there is an active SMTP transaction in progress but the network
channel is closed. The Exception then causes the operationComplete listener
to be called with the appropriate results.